more funcs, regroup elements
