+++ sets/errors.go
@@ -1,0 +1,8 @@
+package sets
+
+import (
+	"errors"
+)
+
+// ErrEmpty is an error for empty set when it's expected to have elements
+var ErrEmpty = errors.New("container is empty")
