+++ sets/set.go
@@ -1,17 +1,6 @@
 package sets
 
-import (
-	c "github.com/life4/genesis/constraints"
-)
-
-// Z is a zero type used as the map value for sets.
-type Z = struct{}
-
-func Contains[S ~map[K]Z, K comparable](set S, item K) bool {
-	var found bool
-	_, found = set[item]
-	return found
-}
+// This file contains functions returning a set.
 
 // Copy returns a shallow copy of the given set.
 func Copy[S ~map[K]Z, K comparable](set S) S {
@@ -20,20 +9,6 @@ func Copy[S ~map[K]Z, K comparable](set S) S {
 		result[k] = Z{}
 	}
 	return result
-}
-
-// Equal returns true if both given sets have the same elements.
-func Equal[S1, S2 ~map[K]Z, K comparable](first S1, second S2) bool {
-	if len(first) != len(second) {
-		return false
-	}
-	for k := range second {
-		_, found := first[k]
-		if !found {
-			return false
-		}
-	}
-	return true
 }
 
 // FromSlice converts a slice into a set.
@@ -54,10 +29,27 @@ func Map[S ~map[K]Z, K, R comparable](set S, f func(K) R) map[R]Z {
 	return result
 }
 
-func Sum[S ~map[K]Z, K c.Integer | c.Float](set S) K {
-	var result K
-	for item := range set {
-		result += item
+// Union returns a set containing all elements from all the given sets.
+func UnionMany[S ~map[K]Z, K comparable](sets ...S) map[K]Z {
+	result := make(map[K]Z)
+	for _, set := range sets {
+		for k := range set {
+			result[k] = Z{}
+		}
 	}
 	return result
 }
+
+// Union
+// Intersection
+// Difference
+// SymmetricDifference
+// EqualMany
+// Filter
+// Reduce
+
+// Add
+// Discard
+// Pop
+// Update
+// Clear
